split test classes
