+++ solace-spring-cloud-stream-binder/solace-spring-cloud-stream-binder/src/test/java/com/solace/spring/cloud/stream/binder/test/util/IgnoreInheritedTests.java
@@ -1,0 +1,17 @@
+package com.solace.spring.cloud.stream.binder.test.util;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * When the test class is marked with {@link IgnoreInheritedTests @IgnoreInheritedTests}
+ * and runs with the {@link InheritedTestsFilteredSpringRunner InheritedTestsFilteredSpringRunner},
+ * all inherited test cases will be filtered from the test run.
+ */
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface IgnoreInheritedTests {
+}
